requirements.txt
flask
transformers
openai-whisper
torch
xformers
